CourseWork corrected

+++ CourseWork/CourseWorkDLL/AirTransport.cpp
@@ -3,7 +3,7 @@
 
 namespace ak
 {
-	float AirTransport::getTravelTime(int distance) const
+	double AirTransport::getTravelTime(int distance) const
 	{
 		return 0.0f;
 	}
