+++ CourseWork/CourseWorkDLL/AllTerrainBoots.cpp
@@ -3,15 +3,15 @@
 
 namespace ak
 {
-	float AllTerrainBoots::getTravelTime(int distance) const
+	double AllTerrainBoots::getTravelTime(int distance) const
 	{
 		int velocity{ 6 };
 		int travelTimeToHalt{ 60 };
 
-		float pureTravelTime = static_cast<float>(distance) / velocity;
-		int haltNum = pureTravelTime / travelTimeToHalt;
+		double pureTravelTime = static_cast<double>(distance / velocity);
+		int haltNum = static_cast<int>(pureTravelTime / travelTimeToHalt);
 
-		float pureRestTime{};
+		double pureRestTime{};
 		if (haltNum == 1)
 		{
 			pureRestTime = 10.0f;
@@ -21,7 +21,7 @@ namespace ak
 			pureRestTime = 10.0f + 5.0f * (haltNum - 1);
 		}
 
-		float travelTime = pureTravelTime + pureRestTime;
+		double travelTime = pureTravelTime + pureRestTime;
 
 		return travelTime;
 	}
