+++ CourseWork/CourseWorkDLL/AllTerrainBoots.h
@@ -12,7 +12,7 @@ namespace ak
 		AllTerrainBoots() : LandTransport("Ботинки-вездеходы") {}
 		~AllTerrainBoots() = default;
 		Transports getType() const override { return Transports::ALL_TERRAIN_BOOTS; }
-		float getTravelTime(int distance) const override;
+		double getTravelTime(int distance) const override;
 	};
 }
