+++ CourseWork/CourseWorkDLL/Eagle.cpp
@@ -3,13 +3,12 @@
 
 namespace ak
 {
-	float Eagle::getTravelTime(int distance) const
+	double Eagle::getTravelTime(int distance) const
 	{
 		int velocity{ 8 };
 		int distanceReductionPercent{ 6 };
 
-		float travelTime = 0.01 * (100 - distanceReductionPercent) * distance /
-			velocity;
+		double travelTime = 0.01f * (100 - distanceReductionPercent) * distance / velocity;
 
 		return travelTime;
 	}
