+++ CourseWork/CourseWorkDLL/LandTransport.cpp
@@ -3,7 +3,7 @@
 
 namespace ak
 {
-	float LandTransport::getTravelTime(int distance) const
+	double LandTransport::getTravelTime(int distance) const
 	{
 		return 0.0f;
 	}
